Crate cjdns_snode

CJDNS supernode

A supernode is a replicating database of node/link information, it's collected by scanning the network for peers but PLEASE DON'T ENABLE SCANNING, there is another snode scanning and you can simply connect to it's socket and listen for all of the updates sent right to your door.

Snode allows dumping of its internal state using TCP/JSON (replication socket) and it allows querying to get a path between any 2 nodes given by their keys using UDP/Benc.

Setup

Build: $ cargo build --release

Create the config file: $ cp config.example.json ./config.json

Start the node: $ ../target/release/cjdns-snode

Modules

args
Command-line arguments parsing.

config
Config file parsing.

supernode
CJDNS supernode implementation.

utils

Functions

main

run