Trait cjdns_snode::pathsearch::graph::GraphSolver

pub trait GraphSolver<T, W> {
    fn path(&self, from: &T, to: &T) -> Vec<T>;

    fn reverse_path(&self, from: &T, to: &T) -> Vec<T>;

    fn path_search_tree(&self, start: &T) -> PathSearchTree<T>;
}

Path finding functions.

Required methods

fn path(&self, from: &T, to: &T) -> Vec<T>

Find a path from from node to to node.

fn reverse_path(&self, from: &T, to: &T) -> Vec<T>

Find a reverse path from to node to from node.

fn path_search_tree(&self, start: &T) -> PathSearchTree<T>

Build a Path Search Tree from a given node.

Loading content...