Struct cjdns_snode::peer::Peers

pub struct Peers {
    peers: PeerList,
    anns: Mutex<AnnList>,
    msg_id_seq: Seq,
    announce_tx: Sender<Vec<u8>>,
}

Fields
peers: PeerListanns: Mutex<AnnList>msg_id_seq: Seqannounce_tx: Sender<Vec<u8>>
Implementations

impl Peers

pub fn get_info(&self) -> PeersInfo

impl Peers

const DROP_AFTER: Duration

const PING_AFTER: Duration

const PING_CYCLE: Duration

pub async fn ping_task<'_>(&'_ self)

async fn do_pings<'_>(&'_ self)

impl Peers

const VERSION: u64

impl Peers

fn new(ann_tx: Sender<Vec<u8>>) -> Self

Create new instance of Peers + announce sender

pub async fn connect_to<'_>(&'_ self, uri: Uri)

Asynchronously start connecting to the specified peer supernode. If the connection can't be established or closed by the remote side, it will be reconnected automatically after a delay.

TODO never returns (should return after async disconnect() call)

pub async fn accept_incoming_connection<'_>(
&'_ self,
from_ipv6: String,
ws_stream: impl WebSock
) -> Result<(), Error>

pub async fn disconnect_all<'_>(&'_ self)

pub async fn add_ann<'_>(&'_ self, hash: AnnHash, binary: Vec<u8>)

pub fn del_ann(&self, hash: &AnnHash)

async fn incoming<'_>(
&'_ self,
addr: String,
ws_stream: impl WebSock
) -> Result<(), Error>

Handle incoming WebSocket connection and process it until closed. This async fn completes when the connection is closed, so spawn a task for it.

async fn outgoing<'_>(
&'_ self,
addr: String,
ws_stream: impl WebSock
) -> Result<(), Error>

Handle outgoing WebSocket connection and process it until closed. This async fn completes when the connection is closed, so spawn a task for it.

fn create_peer<'a>(
&'a self,
addr: String,
ws_stream: impl WebSock + 'a,
peer_type: PeerType
) -> (Peer, impl Future<Output = Result<(), Error>> + 'a)

fn drop_peer(&self, peer: Peer)

async fn run_websocket<'_>(
&'_ self,
peer: Peer,
ws_stream: impl WebSock,
__arg3: Receiver<Message>,
__arg4: Sender<Vec<u8>>
) -> Result<(), Error>

async fn handle_message<'_, '_>(
&'_ self,
__arg1: Peer,
message: Message,
ann_tx: &'_ mut Sender<Vec<u8>>
) -> Result<(), Error>