Trait cjdns_snode::server::route::dijkstra::GraphSolver

pub(super) trait GraphSolver<T: Eq + Hash, W: Eq + Ord> {
    fn path(&self, from: &T, to: &T) -> Vec<T>;

    fn reverse_path(&self, from: &T, to: &T) -> Vec<T>;
}

Path finding functions.

Required methods

fn path(&self, from: &T, to: &T) -> Vec<T>

Find a path from from node to to node.

fn reverse_path(&self, from: &T, to: &T) -> Vec<T>

Find a reverse path from to node to from node.

Loading content...